Make compinit work well with zsh and completions installed via homebrew
With zsh installed from homebrew it looks for completion scripts in those folders:
* /usr/local/share/zsh/site-functions
* /usr/local/share/zsh

Those folders are owned by $USER:admin with permissions for group to write so homebrew could add more completion scripts into them.

compinit raises an error when folder is writable by group:

> For security reasons compinit also checks if the completion system would use files not owned by root or by the current user, or files in directories that are world- or group-writable or that are not owned by root or by the current user.  If such files or directories are found, compinit will ask if the completion system should really be used.  To avoid these tests and make all files found be used without  asking, use the option -u, and to make compinit silently ignore all insecure files and directories use the option -i.  This security check is skipped entirely when the -C option is given.

In order to fix this I disabled errors, it should be safe because those dotfiles are ment to be used on local dev machines.